requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
openai
gitpython
python-dotenv
flake8
pytest
